+++ src/components/modules/Form/FormBox/FormBox.jsx
@@ -1,19 +1,22 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import { useForm, FormProvider } from 'react-hook-form';
 import { Outlet } from 'react-router-dom';
-import { FormWrap } from './FormBox.styled';
 import { Form } from './Form';
+import { FormWrap } from './FormBox.styled';
+import { BackDrop } from '../BackDrop.styled';
 
 export const FormBox = () => {
   const methods = useForm();
+
   return (
     <FormProvider {...methods}>
-      <FormWrap>
-        <Form methods={methods}>
-          <Outlet />
-        </Form>
-      </FormWrap>
+      <BackDrop>
+        <FormWrap>
+          <Form methods={methods}>
+            <Outlet />
+          </Form>
+        </FormWrap>
+      </BackDrop>
     </FormProvider>
   );
 };
