+++ src/components/modules/Form/FormSell/FormSellStepOne.jsx
@@ -42,23 +42,16 @@
 //   );
 // };
 
-import { useState } from 'react';
 import { useFormContext } from 'react-hook-form';
-import { FormSellStepTwo } from './FormSellStepTwo';
-import { Outlet, useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { Text } from 'components/global/text';
-import { useLocation } from 'react-router-dom';
 import { InputWrap, ButtonWrap, TextTittle } from './FormSell.styled';
-import { Link } from 'react-router-dom';
 
 export const FormSellStepOne = () => {
-  // const [open, setOpen] = useState(false);
   const { register, setValue, getValues, reset } = useFormContext();
-  const location = useLocation();
   const navigate = useNavigate();
   return (
     <>
-      {/* <label for="petName">Tittle of ad</label> */}
       <TextTittle>Add pet</TextTittle>
       <Text>Lorem ipsum dolor sit amet, consectetur Lorem ipsum dolor sit amet, consectetur </Text>
       <InputWrap>
@@ -72,14 +65,7 @@ export const FormSellStepOne = () => {
         <input {...register('petBreed')} placeholder="Type breed" id="petBreed" />
       </InputWrap>
       <ButtonWrap>
-        <button
-          type="button"
-          // onClick={() => {
-          //   setOpen(!open);
-          //   console.log('first');
-          // }}
-          onClick={() => navigate('/addpet/step2')}
-        >
+        <button type="button" onClick={() => navigate('/addpet/step2')}>
           Next
         </button>
         <button type="button" onClick={() => navigate('/')}>
