+++ src/components/modules/Form/FormSell/FormSellStepTwo.jsx
@@ -55,35 +55,30 @@ import { useLocation } from 'react-router-dom';
 
 export const FormSellStepTwo = () => {
   const navigate = useNavigate();
-  //   const location = useLocation();
-
-  const [isShown, setIsShown] = useState(true);
 
   const { register, setValue, getValues, reset } = useFormContext();
   return (
     <>
       {/* <label for="petName">Tittle of ad</label> */}
-      {isShown && (
-        <FormWrap>
-          <Text>Add pet</Text>
-          <InputWrap>
-            <label htmlFor="location">Location</label>
-            <input {...register('location')} placeholder="Type location" id="location" />
-            <label htmlFor="location">Price</label>
-            <input {...register('price')} placeholder="Type price" />
-            <label htmlFor="location">Load the pet’s image:</label>
-            <input {...register('photo')} type="file" />
-            <label htmlFor="comments">Comments:</label>
-            <textarea {...register('comments')} />
-          </InputWrap>
-          <ButtonWrap>
-            <button type="button" onClick={() => navigate('/addpet/step1')}>
-              Back
-            </button>
-            <button type="submit"> Done</button>
-          </ButtonWrap>
-        </FormWrap>
-      )}
+      <FormWrap>
+        <Text>Add pet</Text>
+        <InputWrap>
+          <label htmlFor="location">Location</label>
+          <input {...register('location')} placeholder="Type location" id="location" />
+          <label htmlFor="location">Price</label>
+          <input {...register('price')} placeholder="Type price" />
+          <label htmlFor="location">Load the pet’s image:</label>
+          <input {...register('photo')} type="file" />
+          <label htmlFor="comments">Comments:</label>
+          <textarea {...register('comments')} />
+        </InputWrap>
+        <ButtonWrap>
+          <button type="button" onClick={() => navigate('/addpet/step1')}>
+            Back
+          </button>
+          <button type="submit">Done</button>
+        </ButtonWrap>
+      </FormWrap>
     </>
   );
 };
